# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SessionLayer Control Plane API
  version: 0.1.0
  summary: Contract-first REST surface for the SessionLayer Control Plane.
  description: |
    The **SessionLayer Control Plane API** is the contract-first source of truth
    for the platform's REST surface (Design §13, FR-API-1). The UI is a *client* of
    this contract; server interfaces and typed clients are code-generated from this
    document, and CI fails if generated code drifts from the spec. Postgres is the
    single source of truth; the UI and the API share one write path (D11).

    **Complete contract.** This document is the complete resource surface: the
    config resources (`node-policies`, `rules`, `roles` + `role-bindings`, `cas`,
    `capability-defs`, `jit-policies`, `breakglass-policies`,
    `service-accounts`, `session-limit-policies`, `operator-settings`), the
    runtime resources (`nodes`, `gateways`, `sessions`, `session-leases`,
    `locks`, `jit-requests`, `join-tokens`, `otp`, device-flow, `pins`,
    `breakglass`, `recordings`, `audit-events`), and the operational probes.

    **Config vs runtime.** Config resources carry an `origin` provenance label
    (`api|ui|default`) and an optimistic-concurrency `version`; runtime resources
    (locks, sessions, issuance, audit) are never provenance-labelled. `locks` is an
    API-only runtime resource (Design §13).

    **Conventions.** URI versioning (`/v1`); cursor pagination on collections;
...
    compatibility policy.
...
servers:
...
security:
  - oidcBearer: []
  - clientCredentials: []
  - mtls: []
paths:
...
  /v1/node-policies:
    get:
      operationId: listNodePolicies
      summary: List node policies (admin).
      description: |
        Lists node policies (`config.node_policy`; Design §12A) — desired labels,
        connector kind, and declared host-trust references. Cursor-paginated.
        Platform-RBAC gated (`settings:write`).
      tags:
        - node-policies
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of node policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePolicyPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createNodePolicy
      summary: Create a node policy (admin).
      description: |
        Creates a node policy. A bad connector kind, a non-object `desiredLabels`, or
        a host-trust ref that looks like private material is a `422`. Platform-RBAC
        gated (`settings:write`) + audited.
      tags:
        - node-policies
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodePolicyRequest'
      responses:
        '201':
          description: The node policy was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePolicyResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/node-policies/{nodePolicyId}:
    parameters:
      - $ref: '#/components/parameters/NodePolicyId'
    get:
      operationId: getNodePolicy
      summary: Get a node policy (admin).
      description: Returns one node policy by id. Platform-RBAC gated (`settings:write`).
      tags:
        - node-policies
      responses:
        '200':
          description: The node policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePolicyResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateNodePolicy
      summary: Update a node policy (admin).
      description: |
        Replaces a node policy's mutable fields (the `name` is immutable). A stale
        `version` is a `409`. Platform-RBAC gated (`settings:write`) + audited.
      tags:
        - node-policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNodePolicyRequest'
      responses:
        '200':
          description: The updated node policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePolicyResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteNodePolicy
      summary: Delete a node policy (admin).
      description: Deletes a node policy (idempotent). Platform-RBAC gated (`settings:write`) + audited.
      tags:
        - node-policies
      responses:
        '204':
          description: The node policy was deleted (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/recordings:
    get:
      operationId: listRecordings
      summary: List session recordings (admin).
      description: |
        Lists session recording metadata (`runtime.recording_ref`; Design §12).
        NEVER returns recording bytes — replay/export issue short-lived signed URLs.
        Optional filters narrow by session, node, and identity. Cursor-paginated.
        Platform-RBAC gated (`recording:replay`).
      tags:
        - recordings
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: sessionId
          in: query
          required: false
          description: Filter to one session id.
          schema:
            type: string
            format: uuid
        - name: identity
          in: query
          required: false
          description: Filter to one subject identity.
          schema:
            type: string
        - name: nodeId
          in: query
          required: false
          description: Filter to one node id.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A page of recording metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/recordings/{recordingId}:
    parameters:
      - $ref: '#/components/parameters/RecordingId'
    get:
      operationId: getRecording
      summary: Get session recording metadata (admin).
      description: |
        Returns one recording's metadata (never bytes). Platform-RBAC gated
        (`recording:replay`).
      tags:
        - recordings
      responses:
        '200':
          description: The recording metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteRecording
      summary: Governance-delete a recording (admin, privileged).
      description: |
        Governance-mode erasure escape hatch (Design §12.2/§12.3, FR-AUD-3/6): a
        specifically-privileged, audited role deletes the encrypted recording object
        and marks the metadata pruned. Platform-RBAC gated (`recording:delete`) +
        audited (FR-PADM-3). Refused (`409`) for a **compliance**-mode recording
        (truly un-deletable, object-lock) or one under **legal hold**. Idempotent —
        `204` whether or not the object was still present.
      tags:
        - recordings
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: The recording object was governance-deleted (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/recordings/{recordingId}/replay:
    parameters:
      - $ref: '#/components/parameters/RecordingId'
    post:
      operationId: replayRecording
      summary: Issue a replay signed URL (admin).
      description: |
        Issues a short-lived, single-object SIGNED URL for replaying a recording
        (Design §12.2/§18) — bytes NEVER proxy through the CP and the object stays
        customer-key encrypted (the CP cannot decrypt); the access is itself audited
        (FR-AUD-5/FR-PADM-3). Platform-RBAC gated (`recording:replay`), scopable by
        node-label/user/time (FR-PADM-2).
      tags:
        - recordings
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: A short-lived signed URL for replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUrl'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/recordings/{recordingId}/export:
    parameters:
      - $ref: '#/components/parameters/RecordingId'
    post:
      operationId: exportRecording
      summary: Issue an export signed URL (admin).
      description: |
        Issues a short-lived, single-object SIGNED URL for exporting a recording
        (Design §12.2/§18) — bytes NEVER proxy through the CP and the object stays
        customer-key encrypted (the CP cannot decrypt); itself audited (FR-AUD-5).
        Platform-RBAC gated (`recording:export`), scopable (FR-PADM-2).
      tags:
        - recordings
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: A short-lived signed URL for export.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedUrl'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/recordings/{recordingId}/legal-hold:
    parameters:
      - $ref: '#/components/parameters/RecordingId'
    put:
      operationId: setRecordingLegalHold
      summary: Set or release a recording's legal hold (admin, privileged).
      description: |
        Places or releases a legal hold on a recording (Design §12.3, FR-AUD-6). A
        held recording is exempt from retention pruning AND governance delete in
        either WORM mode. Platform-RBAC gated (`recording:delete`) + audited
        (FR-PADM-3). Idempotent by desired state.
      tags:
        - recordings
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegalHoldRequest'
      responses:
        '200':
          description: The updated recording metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC bearer access via an ID/JWT token validated by the Control Plane
        (Design §5.7, FR-AUTH-17). The ID token is the authentication proof.
    clientCredentials:
      type: oauth2
      description: |
        OAuth 2.0 client-credentials for machine consumers (Design §5.6,
        FR-AUTH-12). `private_key_jwt`/mTLS client authentication is preferred
        over static secrets.
      flows:
        clientCredentials:
          tokenUrl: https://controlplane.sessionlayer.example/oauth2/token
          scopes: {}
    mtls:
      type: mutualTLS
      description: |
        Mutual TLS client-certificate authentication (Design §5.7, FR-AUTH-17),
        also used by the internal CP<->Gateway plane.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque forward-only pagination cursor from a previous page's `nextCursor`.
        Omit for the first page. An unrecognised cursor is a `400`.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items to return in one page (the server clamps to its maximum).
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional client-generated idempotency key scoping a mutating request. A
        retry with the same key, method, and path returns the original response and
        never repeats the side effect; the same key with a different request body is
        a `422`. Keys are retained for a bounded TTL.
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    NodePolicyId:
      name: nodePolicyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RecordingId:
      name: recordingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: An error occurred. The body is an RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      title: RFC 9457 Problem Details
      description: |
        Problem details for HTTP APIs as defined by RFC 9457 (obsoletes RFC 7807).
        Error responses use media type `application/problem+json`.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
      additionalProperties: true
...
    Origin:
      type: string
      title: Config Provenance
      description: |
        Provenance of a config row (Design §13). `api`/`ui` mark which admin surface
        last wrote it; `default` is a seeded/cold-start default. Read-only.
      enum:
        - api
        - ui
        - default
...
    ConnectorKind:
      type: string
      title: Connector Kind
      enum:
        - agent
        - agentless
...
    LabelMap:
      type: object
      title: Label Map
      description: A string->string label map.
      additionalProperties:
        type: string
    # ============ rules (dp_rule) ============
...
    NodePolicyResource:
      type: object
      title: Node Policy
      required:
        - id
        - name
        - desiredLabels
        - connectorKind
        - origin
        - version
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        desiredLabels:
          $ref: '#/components/schemas/LabelMap'
        connectorKind:
          $ref: '#/components/schemas/ConnectorKind'
        hostPinRef:
          type: string
        hostCaRef:
          type: string
        origin:
          $ref: '#/components/schemas/Origin'
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateNodePolicyRequest:
      type: object
      title: Create Node Policy Request
      required:
        - name
        - connectorKind
      properties:
        name:
          type: string
          minLength: 1
        desiredLabels:
          $ref: '#/components/schemas/LabelMap'
        connectorKind:
          $ref: '#/components/schemas/ConnectorKind'
        hostPinRef:
          type: string
        hostCaRef:
          type: string
      additionalProperties: false
    UpdateNodePolicyRequest:
      type: object
      title: Update Node Policy Request
      required:
        - connectorKind
        - version
      properties:
        desiredLabels:
          $ref: '#/components/schemas/LabelMap'
        connectorKind:
          $ref: '#/components/schemas/ConnectorKind'
        hostPinRef:
          type: string
        hostCaRef:
          type: string
        version:
          type: integer
          format: int64
      additionalProperties: false
    NodePolicyPage:
      type: object
      title: Node Policy Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NodePolicyResource'
        nextCursor:
          type: string
      additionalProperties: false
    # ============ capability-defs (capability_def) ============
...
    RecordingResource:
      type: object
      title: Recording
      description: Session recording metadata (Design §12). NEVER carries recording bytes.
      required:
        - id
        - sessionId
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        identity:
          type: string
          description: The session subject identity (the recorded user).
        nodeId:
          type: string
          format: uuid
        format:
          type: string
          description: Capture format (e.g. `asciicast-v2`).
        status:
          type: string
          description: Lifecycle status (`recording|finalized|truncated|failed`).
        wormMode:
          type: string
          description: WORM object-lock mode (`compliance|governance`).
        sizeBytes:
          type: integer
          format: int64
        hashChainHead:
          type: string
          description: The tamper-evidence hash-chain head (hex).
        encryptionKeyRef:
          type: string
          description: A reference to the customer-held encryption key — NEVER key material.
        legalHold:
          type: boolean
        retentionUntil:
          type: string
          format: date-time
        prunedAt:
          type: string
          format: date-time
          description: When the encrypted object was deleted (retention prune or governance delete); metadata is retained.
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    LegalHoldRequest:
      type: object
      title: Legal Hold Request
      description: Places or releases a legal hold on a recording (FR-AUD-6).
      required:
        - held
      properties:
        held:
          type: boolean
          description: True to place a legal hold, false to release it.
        reason:
          type: string
          maxLength: 4096
      additionalProperties: false
    RecordingPage:
      type: object
      title: Recording Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecordingResource'
        nextCursor:
          type: string
      additionalProperties: false
    SignedUrl:
      type: object
      title: Signed URL
      description: A short-lived, single-object signed URL. Bytes never proxy through the CP.
      required:
        - url
        - method
        - expiresAt
      properties:
        url:
          type: string
          format: uri
        method:
          type: string
          description: The HTTP method to use against the signed URL (e.g. `GET`).
        expiresAt:
          type: string
          format: date-time
      additionalProperties: false
...